[t119072] Changes the format of the input <operationplan>
The original implementation was done when we still didn't know the
format frePPLe used for the inbound <operationplan>, and made some
assumptions that were wrong, e.g. that the items and locations would
be similar to those of the outbound. Now those inner items have
been placed as attributes of the <operationplan>, following the
specifications given.