[t119559] Uses a lightweight Node to encode XML
Before we were dealing with streams of strings, that complicated things
when extending the methods. I'm in the process of changing, at least
for the export of <customers>, this, so that we can use trees to
extend the code in a more easy way. This will allow us reduce the amount
of monkey patches, since now everything is redirected to an Odoo method,
not a Python class.